Update README.md
I added pseud-code to the end.
Please do not use this to store passwords. It still needs development, it is only a basic start for a password manager. It is only a 32 bit Mersenne Twister algorithm not 64. It is based on prime numbers that are one less than a power of two. So a brute force attack could possibly go right through this. This is my first full project for school so please do not rely on this!!!!!!!!!!!!